prepare project to make it clonable

+++ app/Console/Commands/DeleteOldEvents.php
@@ -7,44 +7,52 @@ use Illuminate\Console\Command;
 use App\Models\Event;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Schema;
 
 class DeleteOldEvents extends Command
 {
   use FileDeletionTrait;
-    /**
-     * The name and signature of the console command.
-     *
-     * @var string
-     */
-    protected $signature = 'app:delete-old-events';
+  /**
+   * The name and signature of the console command.
+   *
+   * @var string
+   */
+  protected $signature = 'app:delete-old-events';
 
-    /**
-     * The console command description.
-     *
-     * @var string
-     */
-    protected $description = 'Удаляет ивенты, которые завершились более 5 дней назад';
+  /**
+   * The console command description.
+   *
+   * @var string
+   */
+  protected $description = 'Удаляет ивенты, которые завершились более 5 дней назад';
 
-    /**
-     * Execute the console command.
-     */
-    public function handle()
-    {
-        // Получаем текущую дату и время
-        $now = Carbon::now();
+  /**
+   * Execute the console command.
+   */
+  public function handle()
+  {
 
-        // Находим ивенты, завершившиеся более 5 дней назад
-        $events = Event::where('datetime', '<', $now->subDays(5))->get();
-
-        foreach ($events as $event) {
-            // удаление ивента
-            $event->delete();
-            // удаление картинки
-            $this->deleteEventFiles($event->id);
-
-            $this->info("Event with ID {$event->id} has been deleted.");
-        }
-
-        $this->info('Old events deleted successfully.');
+    // Проверяем, существует ли таблица 'events'
+    if (!Schema::hasTable('events')) {
+      $this->info('The "events" table does not exist.');
+      return;
     }
+
+    // Получаем текущую дату и время
+    $now = Carbon::now();
+
+    // Находим ивенты, завершившиеся более 5 дней назад
+    $events = Event::where('datetime', '<', $now->subDays(5))->get();
+
+    foreach ($events as $event) {
+      // удаление ивента
+      $event->delete();
+      // удаление картинки
+      $this->deleteEventFiles($event->id);
+
+      $this->info("Event with ID {$event->id} has been deleted.");
+    }
+
+    $this->info('Old events deleted successfully.');
+  }
 }
